Add failing tests for `DocBlockThrows.ThrowMissing`
The "before" stub has two exceptions that should be documented using `@throws`
annotations - the `RangeException` and the `LogicException` - whereas the other
two (`DomainException` and `RuntimeException`) are thrown from nested scopes and
shouldn't be part of the containing method's docblock.

However, without the fix, this sniff reports four missing `@throws` annotations.