Fix WrongNullable false positive
The `nullable` field returned from SignatureTrait::getMethodSignature
is used by DocBlockParamAllowDefaultValueSniff assuming that it means
whether or not the parameter accepts null, so we need to also check if
an explicit null is part of the typehint. This is because in the arrays
returned by File::getMethodParameters, the `nullable_type` field only
indicates if the typehint is prefixed with "?".

Fixes #368.